third_party/WebKit/LayoutTests/http/tests/csspaint/supports-secure-context.html

<!DOCTYPE html>
<script src="../resources/get-host-info.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
  if (window.location.origin != get_host_info().AUTHENTICATED_ORIGIN) {
    window.location = get_host_info().AUTHENTICATED_ORIGIN + window.location.pathname;
  } else {
    test(function() {
      assert_true(CSS.supports('background', 'paint(foo)'));
      assert_true(CSS.supports('content', 'paint(foo)'));
      //assert_true(CSS.supports('border-source-image: paint(foo);'));
    });
  }
</script>
